+++ src/main/java/com/hivemq/mqtt/topic/tree/LocalTopicTree.java
@@ -5,14 +5,12 @@ import com.google.common.collect.ImmutableList;
 import com.google.common.collect.ImmutableMap;
 import com.google.common.collect.ImmutableSet;
 import com.google.common.primitives.ImmutableIntArray;
-import com.hivemq.cluster.topictree.SegmentKeyUtil;
-import com.hivemq.cluster.topictree.TopicSubscribers;
 import com.hivemq.extension.sdk.api.annotations.NotNull;
 import com.hivemq.extension.sdk.api.annotations.Nullable;
 import com.hivemq.mqtt.message.subscribe.Topic;
 import com.hivemq.mqtt.topic.SubscriberWithIdentifiers;
 import com.hivemq.mqtt.topic.SubscriberWithQoS;
-import com.hivemq.persistence.local.xodus.bucket.BucketUtils;
+import com.hivemq.util.BucketUtils;
 import org.apache.commons.lang3.StringUtils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
